chore #38: 패키지 명 소문자로 변경
